camera2: remove fine/normal/superfine jpegquality and use always int
We use now a set of int values the user can exactly assign. Showing as
well the normal fine superfine as choice is just a duplicate. As well
we can get rid of the weird converts of the mixed values.

To take care that the user does not get into problems we updated the db
version and on update we will remove the old value completly due that
it is handled other now. We could start here as well a convert but IMO
overkill to do. So let the user just set a new one.