camera2: set focusmode default always to auto
Idea to set continious focus as default if supported and if not
to fall back to auto is in general not bad. But on some devices
even though continious focus is supported by the camera HAL it is
not really better then auto focus. Prominent example mako and hammerhead.

Whatever change it like it was before to  auto always  as default.
The user is free to change that
in settings later whenever the user want to.